Update packages, move ResourceController into /lib
