+++ lib/ResourceController.js
@@ -12,14 +12,16 @@ var async = require('async');
  * @constructor
  */
 
-function ResourceController(opts) {
+function ResourceController(options) {
     BaseController.call(this);
 
-    if (!opts.model) {
+    options = options || {};
+
+    if (!options.model) {
         throw new Error("'model' property must be defined in 'options' parameter")
     }
 
-    this.model = opts.model;
+    this.model = options.model;
 
     this.hooks = {
         /**
@@ -85,7 +87,8 @@ function ResourceController(opts) {
             get: [],
             getAll: [],
             update: [],
-            delete: []
+            delete: [],
+            any: []
         }
     };
 }
@@ -176,8 +179,6 @@ ResourceController.prototype.compileHandlers = function(operation) {
 
 ResourceController.prototype.task = function (operation) {
     var handlers = this.compileHandlers(operation);
-
-    console.log(handlers);
 
     return function __ResourceController_executeTask(request, response, next) {
         async.eachSeries(handlers, function(listener, callback) {
