+++ package.json
@@ -10,7 +10,7 @@
     "codeclimate-test-reporter": "codeclimate-test-reporter < ./coverage/lcov.info"
   },
   "dependencies": {
-    "@onehilltech/blueprint": "^0.48.1",
+    "@onehilltech/blueprint": "^1.1.1",
     "@onehilltech/blueprint-mongodb": "^0.9.1",
     "bcrypt-nodejs": "^0.0.3",
     "chai": "^3.5.0",
@@ -23,12 +23,13 @@
     "supertest": "^2.0.0",
     "uuid": "^2.0.3",
     "validator": "^6.0.0",
-    "winston": "latest",
-    "lodash": "latest",
-    "mongoose": "latest"
+    "winston": "^2.2.0",
+    "lodash": "^4.16.5",
+    "mongoose": "^4.6.2",
+    "async": "^2.1.2"
+
   },
   "devDependencies": {
-    "async": "latest",
     "chai": "^3.5.0",
     "codeclimate-test-reporter": "^0.4.0",
     "cross-env": "^3.1.3",
